+++ posthog/api/query.py
@@ -20,7 +20,7 @@ from posthog.api.documentation import extend_schema
 from posthog.api.routing import StructuredViewSetMixin
 from posthog.clickhouse.query_tagging import tag_queries
 from posthog.cloud_utils import is_cloud
-from posthog.hogql.database import create_hogql_database, serialize_database
+from posthog.hogql.database.database import create_hogql_database, serialize_database
 from posthog.hogql.errors import HogQLException
 from posthog.hogql.query import execute_hogql_query
 from posthog.models import Team, User
