+++ posthog/hogql/database/fields.py
@@ -1,0 +1,32 @@
+from pydantic import BaseModel, Extra
+from typing import Optional
+
+
+class DatabaseField(BaseModel):
+    """Base class for a field in a database table."""
+
+    class Config:
+        extra = Extra.forbid
+
+    name: str
+    array: Optional[bool]
+
+
+class IntegerDatabaseField(DatabaseField):
+    pass
+
+
+class StringDatabaseField(DatabaseField):
+    pass
+
+
+class StringJSONDatabaseField(DatabaseField):
+    pass
+
+
+class DateTimeDatabaseField(DatabaseField):
+    pass
+
+
+class BooleanDatabaseField(DatabaseField):
+    pass
