+++ posthog/hogql/hogql.py
@@ -2,7 +2,7 @@ from typing import Literal, cast
 
 from posthog.hogql import ast
 from posthog.hogql.context import HogQLContext
-from posthog.hogql.database import create_hogql_database
+from posthog.hogql.database.database import create_hogql_database
 from posthog.hogql.errors import HogQLException, NotImplementedException, SyntaxException
 from posthog.hogql.parser import parse_expr
 from posthog.hogql.printer import prepare_ast_for_printing, print_prepared_ast
