+++ posthog/hogql/resolver.py
@@ -4,7 +4,7 @@ from uuid import UUID
 
 from posthog.hogql import ast
 from posthog.hogql.ast import FieldTraverserType, ConstantType
-from posthog.hogql.database import Database
+from posthog.hogql.database.database import Database
 from posthog.hogql.errors import ResolverException
 from posthog.hogql.visitor import TraversingVisitor
 from posthog.models.utils import UUIDT
