+++ posthog/hogql/test/test_resolver.py
@@ -4,7 +4,7 @@ from uuid import UUID
 from freezegun import freeze_time
 
 from posthog.hogql import ast
-from posthog.hogql.database import create_hogql_database
+from posthog.hogql.database.database import create_hogql_database
 from posthog.hogql.parser import parse_select
 from posthog.hogql.resolver import ResolverException, resolve_types
 from posthog.test.base import BaseTest
